requirements.txt
# Core dependencies
dagster==1.5.0
dagster-webserver==1.5.0
pandas==2.1.0
numpy==1.24.0
pyarrow==14.0.0

# Data sources
aiohttp==3.9.0
requests==2.31.0
jpholiday==0.1.9

# Storage
s3fs==2023.12.0
clickhouse-driver==0.2.6
redis==5.0.0

# Validation
pandera==0.17.0
great-expectations==0.18.0
pydantic==2.5.0

# Feature calculation
pandas-ta==0.3.14b0
scipy==1.11.0

# Testing
pytest==7.4.0
pytest-cov==4.1.0
pytest-asyncio==0.21.0
hypothesis==6.90.0

# Monitoring
prometheus-client==0.19.0

# ML (for later integration)
feast==0.35.0
mlflow==2.9.0

# Utilities
python-dotenv==1.0.0
PyYAML==6.0.1
loguru==0.7.2

# ATFT-GAT-FAN 学習機能依存関係
hydra-core>=1.3.0
torch_geometric>=2.3.0
pywt>=1.4.0  # PyWavelets
mlflow>=2.0.0
optuna>=3.0.0

# ベストプラクティス実装用依存関係
psutil>=5.9.0  # システムメトリクス
matplotlib>=3.7.0  # 可視化
seaborn>=0.12.0  # 統計可視化
polars>=0.19.0  # 高速データ処理
